@@ -1,21 +1,24 @@
 apply plugin: 'com.android.application'
 apply plugin: 'com.google.gms.google-services'
-apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'com.google.firebase.crashlytics'
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 30
     buildToolsVersion "29.0.3"
+
     defaultConfig {
         applicationId "com.gadgetsfolk.kidsgk"
         minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 5
-        versionName "1.01"
+        targetSdkVersion 30
+        versionCode 15
+        versionName "1.1.5"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         android.defaultConfig.vectorDrawables.useSupportLibrary = true
         multiDexEnabled true
     }
+
     buildTypes {
 
         debug {
@@ -34,32 +37,40 @@ android {
         targetCompatibility JavaVersion.VERSION_1_8
     }
 
+    repositories {
+        jcenter()
+        maven {
+            url  "https://adcolony.bintray.com/AdColony"
+        }
+    }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.1'
     //Firebase SDK for Google Analytics
-    implementation 'com.google.firebase:firebase-analytics:17.3.0'
+    implementation 'com.google.firebase:firebase-analytics:17.5.0'
     //Firebase Firestore SDK
-    implementation 'com.google.firebase:firebase-firestore:21.4.2'
+    implementation 'com.google.firebase:firebase-firestore:21.6.0'
     //Multidex Support Library
     implementation 'com.android.support:multidex:1.0.3'
     //Android Design Support Library
-    implementation 'com.google.android.material:material:1.2.0-alpha06'
+    implementation 'com.google.android.material:material:1.3.0-alpha02'
 
     //Rounded ImageView Library
     implementation 'com.rishabhharit.roundedimageview:RoundedImageView:0.8.4'
     //Kotlin Library
-    implementation "androidx.core:core-ktx:1.2.0"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.70"
+    implementation "androidx.core:core-ktx:1.3.1"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.4.0"
     //Firebase Ads SDK
-    implementation 'com.google.firebase:firebase-ads:19.1.0'
-
+    implementation 'com.google.firebase:firebase-ads:19.3.0'
+    // Add the Firebase Crashlytics SDK.
+    implementation 'com.google.firebase:firebase-crashlytics:17.2.1'
 
     implementation 'com.jakewharton:butterknife:10.2.1'
     implementation project(path: ':YouTubeAndroidPlayerApi')
+    //implementation project(':nativetemplates')
     annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.1'
 
     //Glide Library to load images in app.
@@ -67,8 +78,12 @@ dependencies {
     //Glide v4 uses this new annotation processor -- see https://bumptech.github.io/glide/doc/generatedapi.html
     annotationProcessor 'com.github.bumptech.glide:compiler:4.9.0'
 
+    implementation 'com.google.ads.mediation:adcolony:4.1.4.1'
+    implementation 'com.google.ads.mediation:unity:3.4.8.0'
+    implementation 'com.adcolony:sdk:4.2.2'
+    //implementation 'com.google.android.ads:mediation-test-suite:1.4.0'
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test:runner:1.2.0'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    androidTestImplementation 'androidx.test:runner:1.3.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 }
