@@ -9,43 +9,24 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 import com.gadgetsfolk.kidsgk.R;
+import com.gadgetsfolk.kidsgk.helper.HelperMethods;
 import com.gadgetsfolk.kidsgk.model.QuizType;
+import com.google.android.gms.ads.formats.UnifiedNativeAd;
+
 import java.util.ArrayList;
+import java.util.List;
 
-public class QuizTypeAdapter extends RecyclerView.Adapter<QuizTypeAdapter.ViewHolder> {
+public class QuizTypeAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
+    //A Menu ItemView type
+    private static final int MENU_ITEM_VIEW_TYPE = 0;
+    private static final int UNIFIED_NATIVE_AD_VIEW_TYPE = 1;
+
     private Context mContext;
-    private ArrayList<QuizType> mQuizTypeList;
+    private List<Object> mQuizTypeList;
 
-    public QuizTypeAdapter(Context mContext, ArrayList<QuizType> mQuizTypeList) {
+    public QuizTypeAdapter(Context mContext, List<Object> mQuizTypeList) {
         this.mContext = mContext;
         this.mQuizTypeList = mQuizTypeList;
-    }
-
-    @NonNull
-    @Override
-    public QuizTypeAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(mContext).inflate(R.layout.quiz_category_item, parent, false);
-        return new ViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull QuizTypeAdapter.ViewHolder holder, int position) {
-        QuizType quizType = mQuizTypeList.get(position);
-        holder.tvTitle.setText(quizType.getTitle());
-        holder.tvSubTitle.setText(quizType.getSub_title());
-        holder.tvImage.setText(String.valueOf(quizType.getTitle().charAt(0)));
-
-        if (position % 5 == 0) holder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_two));
-        else if (position % 5 == 1) holder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_three));
-        else if (position % 5 == 2) holder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_four));
-        else if (position % 5 == 3) holder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_five));
-        else holder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient));
-    }
-
-
-    @Override
-    public int getItemCount() {
-        return mQuizTypeList.size();
     }
 
     static class ViewHolder extends RecyclerView.ViewHolder{
@@ -61,7 +42,63 @@ public class QuizTypeAdapter extends RecyclerView.Adapter<QuizTypeAdapter.ViewHo
         }
     }
 
-    public void setItems(ArrayList<QuizType> mQuizTypeList){
+    @Override
+    public int getItemCount() {
+        return mQuizTypeList.size();
+    }
+
+    @Override
+    public int getItemViewType(int position) {
+        Object recyclerViewItem = mQuizTypeList.get(position);
+        if (recyclerViewItem instanceof UnifiedNativeAd) {
+            return UNIFIED_NATIVE_AD_VIEW_TYPE;
+        }
+        return MENU_ITEM_VIEW_TYPE;
+    }
+
+    @NonNull
+    @Override
+    public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        switch (viewType){
+            case UNIFIED_NATIVE_AD_VIEW_TYPE:
+                View unifiedNativeLayoutView = LayoutInflater.from(mContext).inflate(R.layout.ad_unified_new, parent, false);
+                return new UnifiedNativeAdViewHolder(unifiedNativeLayoutView);
+            case MENU_ITEM_VIEW_TYPE:
+                //Fall Through
+            default:
+                View view = LayoutInflater.from(mContext).inflate(R.layout.quiz_category_item, parent, false);
+                return new ViewHolder(view);
+        }
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position) {
+        int viewType = getItemViewType(position);
+        switch (viewType) {
+            case UNIFIED_NATIVE_AD_VIEW_TYPE:
+                UnifiedNativeAd nativeAd = (UnifiedNativeAd) mQuizTypeList.get(position);
+                HelperMethods.populateAdView(nativeAd, ((UnifiedNativeAdViewHolder) holder).getView());
+                break;
+            case MENU_ITEM_VIEW_TYPE:
+                //Fall Through
+            default:
+                ViewHolder viewHolder = (ViewHolder) holder;
+
+                QuizType quizType = (QuizType) mQuizTypeList.get(position);
+                viewHolder.tvTitle.setText(quizType.getTitle());
+                viewHolder.tvSubTitle.setText(quizType.getSub_title());
+                viewHolder.tvImage.setText(String.valueOf(quizType.getTitle().charAt(0)));
+
+                if (position % 5 == 0) viewHolder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_two));
+                else if (position % 5 == 1) viewHolder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_three));
+                else if (position % 5 == 2) viewHolder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_four));
+                else if (position % 5 == 3) viewHolder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient_five));
+                else viewHolder.relativeLayout.setBackground(mContext.getResources().getDrawable(R.drawable.gradient));
+        }
+
+    }
+
+    public void setItems(List<Object> mQuizTypeList){
         this.mQuizTypeList = mQuizTypeList;
     }
 }
